saving new snippet is working now

+++ app/src/main/java/xyz/qn0x/copypasta/persistence/entities/SnippetTags.java
@@ -14,29 +14,29 @@ import android.arch.persistence.room.PrimaryKey;
 public class SnippetTags {
 
     @ColumnInfo(name = "snippet_id")
-    private int snippet_id;
+    private long snippet_id;
 
     @ColumnInfo(name = "tag_id")
-    private int tag_id;
+    private long tag_id;
 
-    public SnippetTags(int snippet_id, int tag_id) {
+    public SnippetTags(long snippet_id, long tag_id) {
         this.snippet_id = snippet_id;
         this.tag_id = tag_id;
     }
 
-    public int getSnippet_id() {
+    public long getSnippet_id() {
         return snippet_id;
     }
 
-    public int getTag_id() {
+    public long getTag_id() {
         return tag_id;
     }
 
-    public void setSnippet_id(int snippet_id) {
+    public void setSnippet_id(long snippet_id) {
         this.snippet_id = snippet_id;
     }
 
-    public void setTag_id(int tag_id) {
+    public void setTag_id(long tag_id) {
         this.tag_id = tag_id;
     }
 }
